// cut-by-voice/src/api-router/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartBucket, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  logger: Logger;
  tracer: Tracer;
  VIDEO_PROCESSOR: ServiceStub<import('../video-processor/index.js').default>;
  VIDEO_STORAGE: SmartBucket;
}